Refactor: useImageStore 분리
