+++ src/components/ImageUpload.tsx
@@ -1,15 +1,5 @@
 import React, { useRef } from 'react';
-import { create } from 'zustand';
-
-type StoreState = {
-  uploadedImage: string | null;
-  setUploadedImage: (image: string | null) => void;
-};
-
-const useImageStore = create<StoreState>((set) => ({
-  uploadedImage: null,
-  setUploadedImage: (image) => set({ uploadedImage: image }),
-}));
+import useImageStore from './useImageStore';
 
 const ImageUpload = () => {
   const { uploadedImage, setUploadedImage } = useImageStore();
