+++ src/components/useImageStore.tsx
@@ -1,0 +1,13 @@
+import { create } from 'zustand';
+
+type StoreState = {
+  uploadedImage: string | null;
+  setUploadedImage: (image: string | null) => void;
+};
+
+const useImageStore = create<StoreState>((set) => ({
+  uploadedImage: null,
+  setUploadedImage: (image) => set({ uploadedImage: image }),
+}));
+
+export default useImageStore;
